Broken up DCC script so that Sprog is in it's own class. Clears up code better
